+++ src/shared/components/palette/case-flag/components/add-comments/add-comments.component.spec.ts
@@ -1,12 +1,13 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormGroup, ReactiveFormsModule } from '@angular/forms';
-import { AddCommentsErrorMessage } from '../../enums';
+import { AddCommentsErrorMessage, CaseFlagFieldState } from '../../enums';
 import { AddCommentsComponent } from './add-comments.component';
 
 describe('AddCommentsComponent', () => {
   let component: AddCommentsComponent;
   let fixture: ComponentFixture<AddCommentsComponent>;
+  let nextButton: any;
   let textareaInput: string;
 
   beforeEach(async(() => {
@@ -22,6 +23,7 @@ describe('AddCommentsComponent', () => {
     fixture = TestBed.createComponent(AddCommentsComponent);
     component = fixture.componentInstance;
     component.formGroup = new FormGroup({});
+    nextButton = fixture.debugElement.nativeElement.querySelector('button[type="button"]');
     // 200-character text input
     textareaInput = '0000000000' + '1111111111' + '2222222222' + '3333333333' + '4444444444' + '5555555555' + '6666666666' +
       '7777777777' + '8888888888' + '9999999999' + '0000000000' + '1111111111' + '2222222222' + '3333333333' + '4444444444' +
@@ -33,9 +35,16 @@ describe('AddCommentsComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  it('should show an error message if comments are mandatory but none have been entered', () => {
-    component.validateFlagComments();
+  it('should show an error message on clicking "Next" if comments are mandatory but none have been entered', () => {
+    spyOn(component, 'onNext').and.callThrough();
+    spyOn(component.caseFlagStateEmitter, 'emit');
+    nextButton.click();
     fixture.detectChanges();
+    expect(component.onNext).toHaveBeenCalled();
+    expect(component.caseFlagStateEmitter.emit).toHaveBeenCalledWith({
+      currentCaseFlagFieldState: CaseFlagFieldState.FLAG_COMMENTS,
+      errorMessages: component.errorMessages
+    });
     expect(component.errorMessages[0]).toEqual({
       title: '',
       description: AddCommentsErrorMessage.FLAG_COMMENTS_NOT_ENTERED,
@@ -45,21 +54,28 @@ describe('AddCommentsComponent', () => {
     expect(errorMessageElement.textContent).toContain(AddCommentsErrorMessage.FLAG_COMMENTS_NOT_ENTERED);
   });
 
-  it('should not show an error message if comments are not mandatory and none have been entered', () => {
+  it('should not show an error message on clicking "Next" if comments are not mandatory and none have been entered', () => {
+    spyOn(component, 'onNext').and.callThrough();
+    spyOn(component.caseFlagStateEmitter, 'emit');
     component.optional = true;
     component.ngOnInit();
-    component.validateFlagComments();
+    nextButton.click();
     fixture.detectChanges();
+    expect(component.onNext).toHaveBeenCalled();
+    expect(component.caseFlagStateEmitter.emit).toHaveBeenCalledWith({
+      currentCaseFlagFieldState: CaseFlagFieldState.FLAG_COMMENTS,
+      errorMessages: component.errorMessages
+    });
     expect(component.errorMessages.length).toBe(0);
     const errorMessageElement = fixture.debugElement.nativeElement.querySelector('.govuk-error-message');
     expect(errorMessageElement).toBeNull();
   });
 
-  it('should show an error message if comments exceed a 200-character limit, regardless of optionality', () => {
+  it('should show an error message on clicking "Next" if comments exceed a 200-character limit, regardless of optionality', () => {
     const textarea = fixture.debugElement.nativeElement.querySelector('.govuk-textarea');
     textarea.value = textareaInput + '0';
     textarea.dispatchEvent(new Event('input'));
-    component.validateFlagComments();
+    nextButton.click();
     fixture.detectChanges();
     expect(component.errorMessages[0]).toEqual({
       title: '',
@@ -73,7 +89,7 @@ describe('AddCommentsComponent', () => {
     component.ngOnInit();
     textarea.value = textareaInput + '0';
     textarea.dispatchEvent(new Event('input'));
-    component.validateFlagComments();
+    nextButton.click();
     fixture.detectChanges();
     expect(component.errorMessages[0]).toEqual({
       title: '',
@@ -88,7 +104,7 @@ describe('AddCommentsComponent', () => {
     const textarea = fixture.debugElement.nativeElement.querySelector('.govuk-textarea');
     textarea.value = textareaInput;
     textarea.dispatchEvent(new Event('input'));
-    component.validateFlagComments();
+    nextButton.click();
     fixture.detectChanges();
     expect(component.errorMessages.length).toBe(0);
     let errorMessageElement = fixture.debugElement.nativeElement.querySelector('.govuk-error-message');
@@ -98,7 +114,7 @@ describe('AddCommentsComponent', () => {
     component.ngOnInit();
     textarea.value = textareaInput;
     textarea.dispatchEvent(new Event('input'));
-    component.validateFlagComments();
+    nextButton.click();
     fixture.detectChanges();
     expect(component.errorMessages.length).toBe(0);
     errorMessageElement = fixture.debugElement.nativeElement.querySelector('.govuk-error-message');