+++ src/shared/components/palette/case-flag/write-case-flag-field.component.spec.ts
@@ -1,16 +1,16 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
+import { FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { ActivatedRoute } from '@angular/router';
 import { CaseField } from '../../../domain/definition';
-import { AddCommentsComponent } from './components';
 import { CaseFlagFieldState, CaseFlagStatus } from './enums';
 import { WriteCaseFlagFieldComponent } from './write-case-flag-field.component';
+
+import createSpyObj = jasmine.createSpyObj;
 
 describe('WriteCaseFlagFieldComponent', () => {
   let component: WriteCaseFlagFieldComponent;
   let fixture: ComponentFixture<WriteCaseFlagFieldComponent>;
-  let addCommentsComponent: AddCommentsComponent;
   const flaglauncher_id = 'FlagLauncher';
   const flagLauncherCaseField: CaseField = {
     id: 'FlagLauncher1',
@@ -76,6 +76,7 @@ describe('WriteCaseFlagFieldComponent', () => {
     flagCode: 'WCA',
     status: CaseFlagStatus.INACTIVE
   };
+  const caseFlagsFieldId = 'caseFlags';
   const mockRoute = {
     snapshot: {
       data: {
@@ -96,8 +97,7 @@ describe('WriteCaseFlagFieldComponent', () => {
             {
               id: caseFlag1FieldId,
               field_type: {
-                // TODO: Temporary field type; needs to be changed to "Flags" once the implementation has been changed over
-                id: 'CaseFlag',
+                id: 'Flags',
                 type: 'Complex'
               },
               value: {
@@ -118,8 +118,7 @@ describe('WriteCaseFlagFieldComponent', () => {
             {
               id: caseFlag2FieldId,
               field_type: {
-                // TODO: Temporary field type; needs to be changed to "Flags" once the implementation has been changed over
-                id: 'CaseFlag',
+                id: 'Flags',
                 type: 'Complex'
               },
               value: {
@@ -136,14 +135,21 @@ describe('WriteCaseFlagFieldComponent', () => {
                   }
                 ]
               }
+            },
+            {
+              id: caseFlagsFieldId,
+              field_type: {
+                id: 'Flags',
+                type: 'Complex'
+              },
+              value: {}
             }
           ]
         }
       }
     }
   };
-
-  addCommentsComponent = new AddCommentsComponent();
+  const parentFormGroup = new FormGroup({});
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
@@ -151,8 +157,7 @@ describe('WriteCaseFlagFieldComponent', () => {
       schemas: [ CUSTOM_ELEMENTS_SCHEMA ],
       declarations: [ WriteCaseFlagFieldComponent ],
       providers: [
-        { provide: ActivatedRoute, useValue: mockRoute },
-        { provide: AddCommentsComponent, useValue: addCommentsComponent }
+        { provide: ActivatedRoute, useValue: mockRoute }
       ]
     })
     .compileComponents();
@@ -161,7 +166,9 @@ describe('WriteCaseFlagFieldComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(WriteCaseFlagFieldComponent);
     component = fixture.componentInstance;
-    component.addCommentsComponent = addCommentsComponent;
+    component.caseEditPageComponent = createSpyObj('caseEditPageComponent', ['submit']);
+    component.formGroup = parentFormGroup;
+    component.caseField = flagLauncherCaseField;
     fixture.detectChanges();
   });
 
@@ -179,6 +186,7 @@ describe('WriteCaseFlagFieldComponent', () => {
     // Initial validity of the form is expected to be false because it is at the starting state (only valid at the final state)
     expect(component.isAtFinalState()).toBe(false);
     expect(component.formGroup.valid).toBe(false);
+    expect(component.formGroup.errors).not.toBeNull();
   });
 
   // TODO: Need to re-visit later as the next button has been moved to the child components
@@ -206,7 +214,7 @@ describe('WriteCaseFlagFieldComponent', () => {
     component.caseField = flagLauncherCaseField;
     component.ngOnInit();
     expect(component.flagsData).toBeTruthy();
-    expect(component.flagsData.length).toBe(2);
+    expect(component.flagsData.length).toBe(3);
     expect(component.flagsData[0].flagsCaseFieldId).toEqual(caseFlag1FieldId);
     expect(component.flagsData[0].partyName).toEqual(caseFlag1PartyName);
     expect(component.flagsData[0].roleOnCase).toEqual(caseFlag1RoleOnCase);
@@ -223,31 +231,31 @@ describe('WriteCaseFlagFieldComponent', () => {
     expect(component.flagsData[1].details[1].dateTimeModified).toEqual(new Date(caseFlag1DetailsValue1.dateTimeModified));
     expect(component.flagsData[1].details[1].dateTimeCreated).toEqual(new Date(caseFlag1DetailsValue1.dateTimeCreated));
     expect(component.flagsData[1].details[1].hearingRelevant).toBe(true);
-  });
-
-  // TODO: The below test will be looked at during the future sprint work for case flags
-  // Setting it to non-runnable for now
-  xit('should succeed validate and set flags case field value', () => {
-    spyOn(component.addCommentsComponent, 'validateFlagComments');
-    spyOn(component, 'populateNewFlagDetailInstance').and.returnValue({});
-    component.fieldState = CaseFlagFieldState.FLAG_COMMENTS;
-    component.caseFlagParentFormGroup = new FormGroup({'caseField': new FormControl('')});
-    component.caseFlagParentFormGroup.get('caseField').setValue('sdffsf');
-    component.validateAndSetFlagsCaseFieldValue();
-    expect(component.addCommentsComponent.validateFlagComments).toHaveBeenCalled();
-    expect(component.populateNewFlagDetailInstance).toHaveBeenCalled();
-  });
-
-  it('should fail validate and set flags case field value', () => {
-    component.addCommentsComponent.errorMessages = [{
-      fieldId: 'flagComments', title: '', description: 'Please enter comments for this flag'
-    }];
-    spyOn(component.addCommentsComponent, 'validateFlagComments');
-    component.fieldState = CaseFlagFieldState.FLAG_COMMENTS;
-    component.validateAndSetFlagsCaseFieldValue();
-    expect(component.errorMessages).toEqual(component.addCommentsComponent.errorMessages);
-    expect(component.addCommentsComponent.validateFlagComments).toHaveBeenCalled();
+    expect(component.flagsData[2].flagsCaseFieldId).toEqual(caseFlagsFieldId);
+    expect(component.flagsData[2].partyName).toBeUndefined();
+    expect(component.flagsData[2].roleOnCase).toBeUndefined();
+    expect(component.flagsData[2].details).toBeNull();
   });
 
   // TODO: Need to add tests for when caseField.value is null and caseField.value.details is null
+
+  it('should remove the existing FlagLauncher control from the parent before re-registering', () => {
+    spyOn(parentFormGroup, 'removeControl').and.callThrough();
+    spyOn(component, 'setFlagsCaseFieldValue');
+    // Check the FlagLauncher component control has been registered to the parent FormGroup, and that it is in an invalid
+    // state (intentionally)
+    const flagLauncherFormGroup = parentFormGroup.get(flagLauncherCaseField.id);
+    expect(flagLauncherFormGroup).toBeTruthy();
+    expect(flagLauncherFormGroup.invalid).toBe(true);
+    expect(flagLauncherFormGroup.errors).not.toBeNull();
+    // Move to the final Case Flags stage to make the FlagLauncher control valid
+    component.moveToFinalReviewStage();
+    expect(flagLauncherFormGroup.invalid).toBe(false);
+    expect(flagLauncherFormGroup.errors).toBeNull();
+    // Set the component's formGroup reference back to the parent FormGroup (it gets reassigned in ngOnInit())
+    component.formGroup = parentFormGroup;
+    // Reload the component
+    component.ngOnInit();
+    expect(parentFormGroup.removeControl).toHaveBeenCalledTimes(1);
+  });
 });